Add summary of samples
This file details the number of adult and juvenile (smolt & parr) fish that were analyzed as potential progeny to address a reviewer question. Note: "NA" indicates that that particular year + life stage category was not relevant to our analysis (descended from brood year outside of our 2008-2016 range) and so were not considered.